Struct spacesim::objects::Universe

pub struct Universe { /* fields omitted */ }

The Universe - Every interaction takes place in the relation to the universe.

Methods

impl Universe

fn new(id: i8) -> Universe

Universe constructor - creates single universe containing 1 celestial.

fn get_id(&self) -> i8

Returns id of the universe.

fn spawn(&mut self)

Spawns a celestial in the origin.

fn get_celestial_count(&self) -> i32

Returns a count of all celestials in the given universe.

fn get_celestial(&self, id: i32) -> &Celestial

Returns reference to the celestial with given id.

fn set_position(&mut self, id: i32, point: Coordinates)

Moves celestial by id to a given position.